+++ src/evaluation/mod.rs
@@ -51,12 +51,12 @@ pub struct EvaluationResult {
 impl EvaluationParameters {
     /// Initializes PST patterns with used by default during search.
     fn set_default_pst_patterns(&mut self) {
-        self.pst_patterns[PAWN as usize] = self.get_pawn_pst_pattern();
-        self.pst_patterns[KNIGHT as usize] = self.get_knight_pst_pattern();
-        self.pst_patterns[BISHOP as usize] = self.get_bishop_pst_pattern();
-        self.pst_patterns[ROOK as usize] = self.get_rook_pst_pattern();
-        self.pst_patterns[QUEEN as usize] = self.get_queen_pst_pattern();
-        self.pst_patterns[KING as usize] = self.get_king_pst_pattern();
+        self.pst_patterns[PAWN] = self.get_pawn_pst_pattern();
+        self.pst_patterns[KNIGHT] = self.get_knight_pst_pattern();
+        self.pst_patterns[BISHOP] = self.get_bishop_pst_pattern();
+        self.pst_patterns[ROOK] = self.get_rook_pst_pattern();
+        self.pst_patterns[QUEEN] = self.get_queen_pst_pattern();
+        self.pst_patterns[KING] = self.get_king_pst_pattern();
     }
 
     /// Recalculates initial material and PST tables.
@@ -64,14 +64,14 @@ impl EvaluationParameters {
         for color in WHITE..=BLACK {
             for piece in PAWN..=KING {
                 for phase in OPENING..=ENDING {
-                    self.pst[color as usize][piece as usize][phase as usize] = self.calculate_pst(color, &self.pst_patterns[piece as usize][phase as usize]);
+                    self.pst[color][piece][phase] = self.calculate_pst(color, &self.pst_patterns[piece][phase]);
                 }
             }
         }
     }
 
     /// Calculates PST table for the specified `color` and `pattern`.
-    fn calculate_pst(&self, color: u8, pattern: &[i16; 64]) -> [i16; 64] {
+    fn calculate_pst(&self, color: usize, pattern: &[i16; 64]) -> [i16; 64] {
         let mut array = [0; 64];
 
         match color {
@@ -94,8 +94,8 @@ impl EvaluationParameters {
     }
 
     /// Gets a PST value for the specified `color`, `piece`, `phase` and `square`.
-    pub fn get_pst_value(&self, color: u8, piece: u8, phase: u8, square: u8) -> i16 {
-        self.pst[color as usize][piece as usize][phase as usize][square as usize] as i16
+    pub fn get_pst_value(&self, color: usize, piece: usize, phase: usize, square: u8) -> i16 {
+        self.pst[color][piece][phase][square as usize] as i16
     }
 }
